Update css mark potential bugs
TODO: Rework services to operate with dbcontext instead services itself when creating entity in database. This is bug because service return service model (not entity model) When creating meme (for example) and use category service it always create new category entity in database instead attach existing category to the meme.